Sum multiple legs per route
Individual routes can have legs defined
by waypoints. In such instances there
will be mutiple legs, each with a distance.

This sums the values for each leg and calulates
text based on the result too.

However, for the text it still needs to handle the
potential for different units (metric/km or imperial/miles).
Defaulting to metric for now.